+++ src/protontricks/data/scripts/wine_launch.sh
@@ -2,8 +2,14 @@
 # Helper script created by Protontricks to run Wine binaries using Steam Runtime
 set -o errexit
 
+function log_debug () {
+    if [[ "$PROTONTRICKS_LOG_LEVEL" != "DEBUG" ]]; then
+        return
+    fi
+}
+
 function log_info () {
-    if [[ "$PROTONTRICKS_LOG_LEVEL" != "INFO" ]]; then
+    if [[ "$PROTONTRICKS_LOG_LEVEL" = "WARNING" ]]; then
         return
     fi