+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Added
  - Show app icons for custom shortcuts in the app selector
+ - Verbose flag can be enabled with `-vv` for additional debug logging
 
 ## [1.10.5] - 2023-09-05
 ### Fixed